+++ core/src/com/unciv/ui/multiplayer/EditMultiplayerGameInfoScreen.kt
@@ -1,6 +1,5 @@
 package com.unciv.ui
 
-import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.graphics.Color
 import com.badlogic.gdx.scenes.scene2d.ui.TextField
 import com.unciv.logic.GameInfoPreview
@@ -9,8 +8,7 @@ import com.unciv.logic.civilization.PlayerType
 import com.unciv.models.translations.tr
 import com.unciv.ui.pickerscreens.PickerScreen
 import com.unciv.ui.utils.*
-import com.unciv.ui.worldscreen.mainmenu.OnlineMultiplayer
-import kotlin.concurrent.thread
+import com.unciv.logic.multiplayer.OnlineMultiplayer
 
 /** Subscreen of MultiplayerScreen to edit and delete saves
 * backScreen is used for getting back to the MultiplayerScreen so it doesn't have to be created over and over again */
